requirements.txt
flask==3.1.0
openai==1.70.0
python-dotenv==1.1.0
gunicorn==21.2.0
flask-cors==4.0.0 